Removes TTL
