--- mcc/ttl/README.md
@@ -1,5 +1,0 @@
-# ttl
-
-## Description
-
-`ttl` was discontinued in favor of [got](../got/)
